Combined backToHome component and moved up.
