+++ src/applications/check-in/day-of/pages/CheckIn/DisplayMultipleAppointments.jsx
@@ -4,7 +4,7 @@ import format from 'date-fns/format';
 
 import AppointmentListItem from '../../../components/AppointmentDisplay/AppointmentListItem';
 import BackButton from '../../../components/BackButton';
-import BackToHome from '../../components/BackToHome';
+import BackToHome from '../../../components/BackToHome';
 import Footer from '../../components/Footer';
 import { useFormRouting } from '../../../hooks/useFormRouting';
