+++ src/applications/check-in/day-of/pages/EmergencyContact.jsx
@@ -5,7 +5,7 @@ import PropTypes from 'prop-types';
 import recordEvent from 'platform/monitoring/record-event';
 import { useFormRouting } from '../../hooks/useFormRouting';
 import BackButton from '../../components/BackButton';
-import BackToHome from '../components/BackToHome';
+import BackToHome from '../../components/BackToHome';
 import { focusElement } from 'platform/utilities/ui';
 import Footer from '../components/Footer';
 import { seeStaffMessageUpdated } from '../../actions/day-of';
