+++ src/applications/check-in/day-of/pages/SeeStaff.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useMemo } from 'react';
 import { useSelector } from 'react-redux';
 import PropTypes from 'prop-types';
-import BackToHome from '../components/BackToHome';
+import BackToHome from '../../components/BackToHome';
 import Footer from '../components/Footer';
 import { focusElement } from 'platform/utilities/ui';
 import BackButton from '../../components/BackButton';
