+++ src/applications/check-in/day-of/pages/ValidateVeteran.jsx
@@ -9,7 +9,7 @@ import { createSetSession } from '../../actions/authentication';
 
 import { useFormRouting } from '../../hooks/useFormRouting';
 
-import BackToHome from '../components/BackToHome';
+import BackToHome from '../../components/BackToHome';
 import Footer from '../components/Footer';
 import ValidateDisplay from '../../components/pages/validate/ValidateDisplay';
