+++ src/applications/check-in/pre-check-in/pages/Demographics/index.jsx
@@ -1,7 +1,7 @@
 import React, { useCallback, useEffect, useMemo } from 'react';
 import { focusElement } from 'platform/utilities/ui';
 import { useDispatch, useSelector } from 'react-redux';
-import BackToHome from '../../components/BackToHome';
+import BackToHome from '../../../components/BackToHome';
 import { useFormRouting } from '../../../hooks/useFormRouting';
 import PropTypes from 'prop-types';
 import Footer from '../../components/Footer';
