+++ src/applications/check-in/pre-check-in/pages/EmergencyContact/index.jsx
@@ -8,7 +8,7 @@ import recordEvent from 'platform/monitoring/record-event';
 import { recordAnswer } from '../../../actions/pre-check-in';
 
 import BackButton from '../../../components/BackButton';
-import BackToHome from '../../components/BackToHome';
+import BackToHome from '../../../components/BackToHome';
 import Footer from '../../components/Footer';
 import EmergencyContactDisplay from '../../../components/pages/emergencyContact/EmergencyContactDisplay';
