+++ src/applications/check-in/pre-check-in/pages/NextOfKin/index.jsx
@@ -10,7 +10,7 @@ import { recordAnswer } from '../../../actions/pre-check-in';
 import { api } from '../../../api/';
 
 import BackButton from '../../../components/BackButton';
-import BackToHome from '../../components/BackToHome';
+import BackToHome from '../../../components/BackToHome';
 import Footer from '../../components/Footer';
 import NextOfKinDisplay from '../../../components/pages/nextOfKin/NextOfKinDisplay';
