+++ src/applications/check-in/pre-check-in/pages/Validate/index.jsx
@@ -7,7 +7,7 @@ import { api } from '../../../api';
 
 import { createSetSession } from '../../../actions/authentication';
 
-import BackToHome from '../../components/BackToHome';
+import BackToHome from '../../../components/BackToHome';
 import ValidateDisplay from '../../../components/pages/validate/ValidateDisplay';
 import Footer from '../../components/Footer';
